Local API Connected
- saved api response to formattedResponse for future use
- updated vital threshold global values
- made oxygen saturation float instead of int

+++ lib/screens/patient/patient_tracking.dart
@@ -304,8 +304,8 @@ class PatientTrackingState extends State<PatientTracking> {
       case 'Oxygen Saturation':
         unitLabel = '%';
         validators.addAll([
-          FormBuilderValidators.integer(),
-          FormBuilderValidators.between(50, 100),
+          FormBuilderValidators.numeric(),
+          FormBuilderValidators.between(50.0, 100.0),
         ]);
         inputFormatters.add(FilteringTextInputFormatter.digitsOnly);
         break;
